Gameconf version, Conditionals, various QoL, Copy/Paste tests. (#45)
Closes #4, Closes #24, and Closes #26. Phew!

<img width="288" height="100" alt="image"
src="https://github.com/user-attachments/assets/711cc593-07ed-405c-b412-95201fb4c63f"
/>

<img width="188" height="109" alt="image"
src="https://github.com/user-attachments/assets/484a149c-4eda-4981-9fe8-d32c38bfa6a6"
/>

<img width="251" height="64" alt="image"
src="https://github.com/user-attachments/assets/44a8273a-a0e0-471e-b24c-0a6cc1465f4d"
/>


---

_Have you heard of "Copy/Paste inputs on Windows"? It's the world's ONLY
regression!_

+++ src/conditions.rs
@@ -239,8 +239,8 @@ fn check_game_state_condition(condition: &ConditionDef, state: &PreviewState) ->
     use crate::model::ConditionType::*;
     let param = condition.param;
     match condition.condition {
-        GameVersionGe => state.world.game_version >= param,
-        GameVersionLt => state.world.game_version < param,
+        GameVersionGe => (state.world.game_version as i32) >= param,
+        GameVersionLt => (state.world.game_version as i32) < param,
         SessionTypeEq => state.world.session_type == param,
         SessionTypeNeq => state.world.session_type != param,
         GameModeEq => true,
